[test] apply the new test style to databend-query - api (46/177)
