+++ query/tests/it/api/rpc/flight_service.rs
@@ -24,14 +24,14 @@ use common_exception::exception::ABORT_SESSION;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_planners::Expression;
+use databend_query::api::DatabendQueryFlightDispatcher;
+use databend_query::api::DatabendQueryFlightService;
+use databend_query::api::FlightAction;
+use databend_query::api::FlightTicket;
+use databend_query::api::ShuffleAction;
+use databend_query::api::StreamTicket;
 use tonic::Request;
 
-use crate::api::rpc::flight_actions::FlightAction;
-use crate::api::rpc::flight_tickets::StreamTicket;
-use crate::api::rpc::DatabendQueryFlightDispatcher;
-use crate::api::rpc::DatabendQueryFlightService;
-use crate::api::FlightTicket;
-use crate::api::ShuffleAction;
 use crate::tests::create_query_context;
 use crate::tests::parse_query;
 use crate::tests::SessionManagerBuilder;
