+++ query/tests/it/api/rpc/mod.rs
@@ -12,23 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use flight_actions::BroadcastAction;
-pub use flight_actions::CancelAction;
-pub use flight_actions::FlightAction;
-pub use flight_actions::ShuffleAction;
-pub use flight_client::FlightClient;
-pub use flight_dispatcher::DatabendQueryFlightDispatcher;
-pub use flight_service::DatabendQueryFlightService;
-pub use flight_tickets::FlightTicket;
-pub use flight_tickets::StreamTicket;
-
 mod flight_actions;
-mod flight_client;
-mod flight_client_stream;
 mod flight_dispatcher;
-mod flight_scatter;
-mod flight_scatter_broadcast;
-mod flight_scatter_hash;
 mod flight_service;
-mod flight_service_stream;
 mod flight_tickets;
